% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_rl_network.R
\name{analyze_rl_network}
\alias{analyze_rl_network}
\title{Identifies all R / L interactions}
\usage{
analyze_rl_network(input, h = 5, w = 20)
}
\arguments{
\item{input}{the input network from plot_rl_network}
}
\description{
This function will map all RL interactions
}
\details{
This will use the calc_agg_bulk results to ID networks
}
\examples{
ex_sc_example <- id_rl(input = ex_sc_example)
}
