% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_agg_bulk.R
\name{calc_agg_bulk}
\alias{calc_agg_bulk}
\title{Calculate UPM expression values across pData values}
\usage{
calc_agg_bulk(input, aggregate_by, group_by = FALSE,
  cutoff_frac = FALSE, cutoff_num = FALSE, cutoff_cpm = FALSE)
}
\arguments{
\item{input}{the input ex_sc.}

\item{aggregate_by}{The pData variables to break by}

\item{group_by}{The pData variables that contains distinct groups if included in aggregate_by (eg Timepoint, Condition, etc)
This is used to calculate portions internally to this group}

\item{cutoff_frac}{a  fraction expressing value below which gene expression values will be set to 0 for the mean.}

\item{cutoff_num}{a total number cells expressing value below which gene expression values will be set to 0 for the mean.}

\item{cutoff_cpm, }{the minimum CPM values below which set to 0
This is useful for removing nodes from networks that contain only a couple of cells.}
}
\description{
This will calculate UMIs per million UPM expression values across pData columns. Useful for heatmaps and networking analysis.
}
\details{
Utilize information stored in pData to control the plot display.
}
\examples{
plot_tsne_metadata(ex_sc_example, color_by = "UMI_sum", title = "UMI_sum across clusters", facet_by = "Cluster", ncol = 3)
}
