% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_cpg.R
\name{calc_cpg}
\alias{calc_cpg}
\title{Calculate the cells per gene}
\usage{
calc_cpg(ex_sc)
}
\arguments{
\item{ex_sc}{an expression set}
}
\description{
This function will calculate how many cells have non-zero expression values of each gene.
}
\details{

}
\examples{
calc_cpg(input_norm)
}
