% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_genestats.R
\name{calc_genestats}
\alias{calc_genestats}
\title{Calculate Gene Stats}
\usage{
calc_genestats(input)
}
\arguments{
\item{input}{the input ex_sc}
}
\description{
This function will calculate basic statistics about gene usage
}
\details{
This will calculate the total number of UMIs on a per cell basis.
}
\examples{
ex_sc_example <- calc_genestats(input = ex_sc_example)
}
