% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_libsize.R
\name{calc_libsize}
\alias{calc_libsize}
\title{Calculate Library Size}
\usage{
calc_libsize(input, suffix = "")
}
\arguments{
\item{input}{the input ex_sc}
}
\description{
This function will count the number of reads / UMIs per cell
}
\details{
This will calculate total UMIs on a per cell basis.
}
\examples{
ex_sc_example <- calc_libsize(input = ex_sc_example)
}
