% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rl_network.R
\name{calc_rl_network}
\alias{calc_rl_network}
\title{Identifies all R / L interactions}
\usage{
calc_rl_network(input, nodes, group_by = FALSE, weight_by_proportion = TRUE,
  print_progress = TRUE)
}
\arguments{
\item{input}{the input ex_sc}

\item{nodes}{the pData variable used in calc_agg_bulk that will be used to place nodes (such as cluster or cell type)}

\item{group_by}{the pData columns calc_agg_bulk was calculated on to split the netoworks into independent networks
(such as an experimental condition)}

\item{weight_by_proportion}{if true will multiple ligand mean expression within a node by the proportion of that group}
}
\description{
This function will map all RL interactions
}
\details{
This will use the calc_agg_bulk results to ID networks
}
\examples{
ex_sc_example <- id_rl(input = ex_sc_example)
}
