% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_sc.R
\name{cluster_sc}
\alias{cluster_sc}
\title{Cluster_sc
This will perform clustering on your single cell data.}
\usage{
cluster_sc(input, method, num_clust)
}
\arguments{
\item{input}{the input data matrix.}

\item{method}{can either be "spectral" or "density"}

\item{num_clust}{the number of clusters}
}
\description{
Cluster_sc
This will perform clustering on your single cell data.
}
\details{
This selects genes.
}
\examples{
clusters <- cluster_sc(input = exprs(ex_sc_example), method = "spectral", num_clust)
}
