% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_ex_sc.R
\name{construct_ex_sc}
\alias{construct_ex_sc}
\title{Construct Expression Set Class}
\usage{
construct_ex_sc(input)
}
\arguments{
\item{input}{the input data matrix.}
}
\description{
This function will take an input expression matrix and make an ESC
}
\details{
This will take an input matrix and create the expression set class that further analysis can be written into.
}
\examples{
construct_ex_sc(filtered_data)
}
