% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_ex_sc_10x.R
\name{construct_ex_sc_10x}
\alias{construct_ex_sc_10x}
\title{Calculate Library Size}
\usage{
construct_ex_sc_10x(bc_path, long_genes = T, feature_assay = T)
}
\arguments{
\item{input}{the input DIRECTORY}
}
\description{
This function will count the number of reads / UMIs per cell
}
\details{
This will calculate the total number of UMIs on a per cell basis.
}
\examples{
ex_sc_example <- calc_libsize(input = ex_sc_example)
}
