% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_table.R
\name{DE_table}
\alias{DE_table}
\title{create table of DE results}
\usage{
DE_table(folders, celltype, input, keeps = c("logFC", "FDR"),
  filter_by = "FDR", number = 30000, fc_cutoff = 1e-06)
}
\arguments{
\item{folders}{storing the output files of findDEgenes(), with cell type as facet_by.}

\item{celltype}{the cell type to select}

\item{input}{the ex_sc used for findDEgenes()}

\item{keeps}{the column in the result table that want to keep}

\item{filter_by}{the column to filter by}

\item{number}{rank the 'filter_by', keep the top 'number'}

\item{fc_cutoff}{used for logFC filter}
}
\description{
This function will create a table with each gene as a row and each DE output file as a column.
}
\details{
Use large 'number' and small 'fc_cutoff' to get all the genes kept.
}
\examples{

}
