% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_reduce.R
\name{dim_reduce}
\alias{dim_reduce}
\title{Initial Dimension Reduction}
\usage{
dim_reduce(input, pre_reduce = "ICA", nComp = 10, tSNE_perp = 30,
  print_progress = TRUE)
}
\arguments{
\item{input}{the input data matrix.}

\item{pre_reduce}{the algorithm choice for reduction before tSNE (either "ICA", "PCA", "iPCA").}

\item{nComp}{the number of components to reduce too before tSNE, 5-20 recommended.}

\item{tSNE_perp}{number of cells expressed above threshold for a given gene, 10-100 recommended.}

\item{print_progress}{Print to the terminal progress information.}
}
\description{
This function will do dimensionality reduction.
}
\details{
The minCells will be highly dependent on the number of cells in your dataset. In general aim to include genes expressed in no less than 1:100 cells, unless you have extremely rare cell types below that threshold.
}
\examples{
filtered_data <- dim_reduce(input = filtered_data, pre_reduce = "ICA", nComp = 10, tSNE_perp = 30, print_progress=TRUE)
}
