% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_density.R
\name{draw_density}
\alias{draw_density}
\title{Draw Density}
\usage{
draw_density(input, val, var1 = "NA", statistic = "mean")
}
\arguments{
\item{input}{the input data}

\item{val}{Either "UMIs" or a gene name}

\item{var1}{The metadata in pData that will be used to break the distributions into groups. Can be left blank.}

\item{statistic}{Either the "mean" or "median" can be calculated}
}
\description{
This function will draw a UMI distribution
}
\details{
This will draw a UMI distribution for each variable in the metadat.
}
\examples{
draw_density(input = ex_sc_example, val = "UMIs", var1 = "Cluster", statistic = "mean")
}
