% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeRDE.R
\name{edgeRDE}
\alias{edgeRDE}
\title{This will run edgeR to find differentially expressed genes. Use the wrapping functions findDEgenes or findDEmarkers to use pData columns and get the proper input}
\usage{
edgeRDE(input, groups, batch, sizefactor, lib_size, minCells, pVal = 1,
  contrast = list(c(-1, 1)))
}
\arguments{
\item{input}{Input expression set}

\item{groups}{an integer vector with the group for each cell}

\item{batch}{a vector with the batch for each cell, if you don't need batches create a vector with the same id for all cells}

\item{sizefactor}{a pData column containing the scran reported size factor for each cell}

\item{lib_size}{a pData column containing the library size for each cell}

\item{minCells}{minimum cell fraction the genes should be expressed in to be tested for DE}

\item{pVal}{pvalue cutoff for reported results, by default reports all results}

\item{contrast}{a list of contrasts to be used for the DE analysis, default is a two-class comparison, with the second class as the comparison class}
}
\description{
This will run edgeR to find differentially expressed genes. Use the wrapping functions findDEgenes or findDEmarkers to use pData columns and get the proper input
}
\details{
Utilize information stored in pData to control the DE performed
The result is a list with the contrast results, access each element for the DE table
}
\examples{
myDEresults = edgeRDE(input, groups = mygroup, batch = beads, sizefactor = "sizefactor", lib_size = "lib_size")
}
