% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_nodes.R
\name{extract_nodes}
\alias{extract_nodes}
\title{Identifies all R / L interactions}
\usage{
extract_nodes(input, nodes, expand = 0)
}
\arguments{
\item{input}{the input graph analysis object}

\item{nodes}{the nodes to extract}

\item{expand}{if true will expand to grab all nodes within distance n of a node of interest}
}
\description{
This function will map all RL interactions
}
\details{
This will use the calc_agg_bulk results to ID networks
}
\examples{
ex_sc_example <- id_rl(input = ex_sc_example)
}
