% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_rl_network.R
\name{filter_rl_network}
\alias{filter_rl_network}
\title{Identifies all R / L interactions}
\usage{
filter_rl_network(input, filter_by, cutoff, direction, group_by = FALSE)
}
\arguments{
\item{input}{the output of calc_rl_network}

\item{filter_by}{the column of full_network to summarize by}

\item{cutoff}{value to filter by}

\item{direction}{>, <, =}
}
\description{
This function will map all RL interactions
}
\details{
This will use the calc_agg_bulk results to ID networks
}
\examples{
ex_sc_example <- id_rl(input = ex_sc_example)
}
