% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_UMIs.R
\name{filter_UMIs}
\alias{filter_UMIs}
\title{This function will filter based on min number of UMIs, max number of UMIs, and genes based on expression across all cells.}
\usage{
filter_UMIs(input, minUMI, maxUMI, threshold, minCells)
}
\arguments{
\item{minUMI}{min number of UMIs per cell}

\item{maxUMI}{max number of UMIs per cell}

\item{threshold}{UMI threshold for gene detection}

\item{minCells}{number of cells expressed above threshold for a given gene}
}
\description{
This function will filter based on min number of UMIs, max number of UMIs, and genes based on expression across all cells.
}
\examples{
filter_UMIs()
}
\keyword{cats}
