% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findDEgenes.R
\name{findDEgenes}
\alias{findDEgenes}
\title{This will perform differential expression using edgeR for pairwise comparisons}
\usage{
findDEgenes(
  input,
  pd = pData(input),
  DEgroup,
  contrastID,
  sizefactor = NULL,
  lib_size = NULL,
  facet_by,
  minCells = 0.1,
  batchID = NULL,
  outdir = "DEresults/",
  outsuffix = "DEresults.tsv",
  pVal = 1,
  contrast = list(c(-1, 1))
)
}
\arguments{
\item{input}{Input expression set}

\item{pd}{pData, optional, by default set to pData of input}

\item{DEgroup}{a pData variable}

\item{contrastID}{the condition value from DEgroup to be used as the contrast}

\item{sizefactor}{a pData column containing the scran reported size factor for each cell}

\item{lib_size}{a pData column containing the library size for each cell}

\item{facet_by}{a pData variable to iterate through and perform DE for a condition within each group}

\item{minCells}{minimum cell fraction the genes should be expressed in to be tested for DE}

\item{batchID}{a pData variable to use for model accounting for batch effects}

\item{outdir}{an output directory}

\item{outsuffix}{an output file suffix}

\item{pVal}{pvalue cutoff for reported results, by default reports all results}

\item{contrast}{a list of contrasts to be used for the DE analysis, default is a two-class comparison, with the second class as the comparison class}
}
\description{
This will perform differential expression using edgeR for pairwise comparisons
}
\details{
Utilize information stored in pData to control the DE performed
}
\examples{
findDEgenes(input, pd = selected_table, DEgroup = "genotype", batchID = "beads", facet_by = "cluster", sizefactor="sizefactor", lib_size="lib_size")
}
