% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findDEmarkers.R
\name{findDEmarkers}
\alias{findDEmarkers}
\title{This will perform differential expression (DE) using edgeR to find markers for each cluster or group compared to all others}
\usage{
findDEmarkers(input, pd = pData(input), DEgroup, batchID, sizefactor,
  lib_size, outdir = "DEmarkers/", outsuffix = "DEmarkers.tsv",
  minCells = 0.1, pVal = 1, contrast = list(c(-1, 1)))
}
\arguments{
\item{input}{Input expression set}

\item{pd}{pData (optional) by default set to pData(input)}

\item{DEgroup}{a pData variable}

\item{batchID}{a pData variable to use for model accounting for batch effects}

\item{sizefactor}{a pData column containing the scran reported size factor for each cell}

\item{lib_size}{a pData column containing the library size for each cell}

\item{outdir}{an output directory (default="DEmarkers/")}

\item{outsuffix}{an output file suffix (default="DEmarkers.tsv")}

\item{minCells}{minimum cell fraction the genes should be expressed in to be tested for DE (default = 10%)}

\item{pVal}{pvalue cutoff for reported results (default=1, reports all results)}

\item{contrast}{a list of contrasts to be used for the DE analysis (default is a two-class comparison, with the second class as the comparison class)}
}
\description{
This will perform differential expression (DE) using edgeR to find markers for each cluster or group compared to all others
}
\details{
Utilize information stored in pData to control the DE performed
}
\examples{
findDEmarkers(input, pd = selected_table, DEgroup = "cluster", batchID = "beads", sizefactor="sizefactor", lib_size="lib_size")
###
}
