% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_filter.R
\name{flow_filter}
\alias{flow_filter}
\title{Construct Expression Set Class}
\usage{
flow_filter(input, panels, title)
}
\arguments{
\item{input}{the input data ex_sc}

\item{panels}{the list of panels to be gated on}

\item{title}{the title of the plot}
}
\description{
This function will take an input expression matrix and make an Expression Set Class
}
\details{
This will take an expression set class and a list of flow like panels and assign cells to groups based on the expression of these genes
}
\examples{
flow_filter(input = sc_dat, panels = panel_list)
}
