% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_svm.R
\name{flow_svm}
\alias{flow_svm}
\title{Flow Support Vector Machine}
\usage{
flow_svm(input, pcnames)
}
\arguments{
\item{input}{the input data ex_sc}
}
\description{
This function will take an expression set class after flow sorting, and train to identify cells that failed to pass the gates.
}
\details{
This will take an expression set class and classify cells via SVM that did not pass the gates.
Note that the SVM parameters are all defaults. You are encourages to modify this function to suit your purposes.
This will do SVM training and classification on the Components that were fed INTO tSNE.
}
\examples{
flow_svm(input = sc_dat)
}
