% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist2.R
\name{hist2}
\alias{hist2}
\title{Plot Frequency Histogram}
\usage{
hist2(data, break_by = 1, barnum = 0, main = "", xlab = "")
}
\arguments{
\item{data}{a numeric vector}

\item{break_by}{the range of each bar}

\item{barnum}{number of bars in total}

\item{main}{title}

\item{xlab}{x-axis label}
}
\description{
This function will calculate the frequency and plot the histogram
}
\details{
This is build on hist(), with a easier way to set the breaks.
}
\examples{

}
