% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_markers.R
\name{id_markers}
\alias{id_markers}
\title{ID markers
This will perform marker gene identification}
\usage{
id_markers(input, num_markers, num_reference)
}
\arguments{
\item{input}{the input data matrix.}

\item{num_markers}{the number of markers to return}
}
\description{
ID markers
This will perform marker gene identification
}
\details{
This will find marker genes for each cluster.
}
\examples{
ex_sc_example <- id_markers(input = exprs(ex_sc_example), num_markers = 50, num_reference = 1000)
}
