% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_rl.R
\name{id_rl}
\alias{id_rl}
\title{identify receptors and ligands}
\usage{
id_rl(input)
}
\arguments{
\item{input}{the input ex_sc}
}
\description{
This function will id receptors and ligands in your data and mark them in fData
}
\details{
This will use data(Receptor_Ligand_Data) to write to fData 2 columns, one for
receptors and one for ligands. TRUE FALSE statements will be written depending on their presence.
Note that this uses HUMAN GENE NAMES!!
}
\examples{
ex_sc_example <- id_rl(input = ex_sc_example)
}
