% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_ex_sc.R
\name{merge_ex_sc}
\alias{merge_ex_sc}
\title{This will merge pData and fData}
\usage{
merge_ex_sc(input, subset_input, merge_key)
}
\arguments{
\item{input}{the expression set to merge into}

\item{subset_input}{the expression set to be merged (contains a SUBSET of input1)}

\item{merge_key}{this key will be appended to the colnames of x,y, and cluster}

\item{include}{defaults to all}
}
\description{
This function will take an input ex_sc and produce a subsetted one preserving all metadata
}
\details{
If you have subsetted data and want to merge the tSNE coordinates and cluster information into the master data, this will do so.
}
\examples{
merged_ex_sc <- merge_ex_sc(input1, input2, merge_key)
}
