% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_sc.R
\name{normalize_sc}
\alias{normalize_sc}
\title{normalize_sc
This will perform normalization on your single cell data.}
\usage{
normalize_sc(input, method)
}
\arguments{
\item{input}{the input data matrix.}

\item{method}{either "CPM", "TMM", "SCRAN", "scNORM"}
}
\description{
normalize_sc
This will perform normalization on your single cell data.
}
\details{
This normalizes data with one of three methods.
}
\examples{
normalized_dat <- normalize_sc(input = exprs(ex_sc_example))
}
