% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_barplot.R
\name{plot_barplot}
\alias{plot_barplot}
\title{This will create a bar plot}
\usage{
plot_barplot(
  input,
  title = "",
  gene,
  color_by,
  facet_by = "NA",
  colors = NA,
  ncol = "NA",
  number_labels = T,
  text_sizes = c(20, 10, 5, 10, 5, 5),
  theme = "classic",
  alpha = 0.5
)
}
\arguments{
\item{input}{The input data}

\item{title}{The title}

\item{gene}{if provided will color_by the gene}

\item{color_by}{a pData variable}

\item{facet_by}{a pData variable}

\item{colors}{What colors to utilize for categorial data. Be sure it is of the proper length!}

\item{ncol}{How many columns if faceting}

\item{text_sizes}{a vector of title_size, axis_title, axis_text, legend_title, legend_text, facet_text, faults too c(20,10,5,10,5,5)}

\item{theme}{the plot theme}
}
\description{
This will plot a given gene via violin plot
}
\details{
Utilize information stored in pData to control the plot display.
}
\examples{
plot_violin(input, title = "Actb across clusters", gene = "Actb", color_by = "Timepoint", facet_by = "Cluster", size = 1, ncol = 3)
}
