% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_density_ridge.R
\name{plot_density_ridge}
\alias{plot_density_ridge}
\title{Plot Density}
\usage{
plot_density_ridge(input, val, title, color_by)
}
\arguments{
\item{input}{the input ex_sc}

\item{val}{Either "UMI_sum" or a gene name}

\item{title}{The title of the plot}

\item{color_by}{The metadata in pData that will be used to break the distributions into groups. Can be left blank.}
}
\description{
This function will draw a distribution for a given value
}
\details{
This will draw a UMI distribution for a given value, broken into groups if color_by != "NA"
}
\examples{
draw_density(input = ex_sc_example, val = "UMI_sum", color_by = "Cluster", statistic = "mean")
}
