% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dotplot.R
\name{plot_dotplot}
\alias{plot_dotplot}
\title{This will create a bar plot with dots}
\usage{
plot_dotplot(input, gene, color_by, facet_by = "NA", point_by,
  title = "", colors = NA, ncol = "NA", number_labels = T,
  text_sizes = c(20, 10, 5, 10, 5, 5), theme = "classic",
  alpha = 0.5, stackratio = 0.4, dotsize = 3, bar = T,
  binwidth = 0.005, link = FALSE)
}
\arguments{
\item{input}{Bioconductor’s ExpressionSet Class with bulk value stored in fData.}

\item{gene}{to plot the expression level of}

\item{color_by}{a pData variable}

\item{facet_by}{a pData variable}

\item{point_by}{a pData variable.}

\item{title}{The title}

\item{colors}{What colors to utilize for categorial data. Be sure it is of the proper length!}

\item{ncol}{How many columns if faceting}

\item{number_labels}{to show the cell numbers and cell percentage of each bar.}

\item{text_sizes}{a vector of title_size, axis_title, axis_text, legend_title, legend_text, facet_text, faults too c(20,10,5,10,5,5)}

\item{theme}{the plot theme}

\item{stackratio}{the overlap of dots.}

\item{dotsize}{the size of dots.}

\item{bar}{the weighted mean.}

\item{binwidth}{average the values when dots are within the range of (max-min)*binwidth.}

\item{link}{link the dots from same patients in each panel.}
}
\description{
This will create a bar plot with dots
}
\details{
Utilize information stored in pData to control the plot display. Each point_by as a dot with a bar showing the weighted mean of all point_by dots.
}
\examples{
plot_dotplot(ex_sc, gene = "ADCY7", color_by = "Skin", facet_by = "subCellType", point_by = "Patient")
}
