% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gene_dots.R
\name{plot_gene_dots}
\alias{plot_gene_dots}
\title{Plot of genes by pData variable}
\usage{
plot_gene_dots(input, genes, break_by, title = "", scale_by = FALSE, log = F)
}
\arguments{
\item{input}{the input ex_sc.}

\item{genes}{What to color points by, either "UMI_sum", or pData categorial variable, ignored if gene is provided}

\item{break_by}{a pData variable to break the x axis by}

\item{title}{The title}

\item{scale_by}{Whether each gene should be scaled by break_by ("col") or gene ("row")}
}
\description{
This will plot information onto a 2d tsne plot
}
\details{
Utilize information stored in pData to control the plot display.
}
\examples{
plot_tsne_metadata(ex_sc_example, color_by = "UMI_sum", title = "UMI_sum across clusters", facet_by = "Cluster", ncol = 3)
}
