% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_go_heatmap.R
\name{plot_go_heatmap}
\alias{plot_go_heatmap}
\title{Construct Expression Set Class}
\usage{
plot_go_heatmap(
  input,
  sig_val = "p.adjust",
  cutoff = 0.1,
  max_categories = 30,
  color_by = "sig",
  subset_group = F,
  prune = F,
  cutoff_jaccard = 1,
  word_similarity = 0.5
)
}
\arguments{
\item{input}{the input Go result from clusterProfiler}

\item{sig_val}{p.adjust of pvalue}

\item{cutoff}{the value to be included}

\item{max_categories}{mox categories included}

\item{color_by}{what the cells should be shaded by. "count" or "sig"}

\item{subset_group}{table(input@compareClusterResult$group)}

\item{prune}{If true will prune the resulting heatmap based on jaccard similarity}

\item{cutoff_jaccard}{The jaccard similarity index of terms to create a new group}

\item{word_similarity}{for groups of go terms, the fraction of the original terms that a given word must occur in}
}
\description{
This function will take an input expression matrix and make an Expression Set Class
}
\details{
This will take the output of a GO enrichment analysis or GSEA object from clusterProfiler
and merge terms based on the overlap between the enriched genes
}
\examples{
plot_go_heatmap(input = gsea_result)
}
