% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plots a heatmap}
\usage{
plot_heatmap(input, genes, type, title = "Heatmap", scale_by = "row",
  cluster_by = "row", color_pal = viridis::magma(256), text_angle = 60,
  group_names = TRUE, gene_names = TRUE, facet_by = FALSE)
}
\arguments{
\item{input}{the input ex_sc}

\item{genes}{a vector of genes to go into the heatmap}

\item{type}{can either be "bulk" or "single_cell. If type is "bulk" it will utilize the bulk info stored in fData}

\item{title}{A heatmap title}

\item{scale_by}{scale across "row" (genes), "col" (groups), or FALSE}

\item{cluster_by}{either "row", col, or both}

\item{color_pal}{The color pallete to be used}

\item{text_angle}{The desired angle for text on the group labels}

\item{group_names}{whether groups should be labelled}

\item{gene_names}{whether genes should be labelled}

\item{facet_by}{will create breaks in the heatmap by some pData Variable}
}
\description{
Makes either a single cell or bulk heatmap,
}
\details{
Utilize information stored in pData to control the plot display.
}
\examples{
plot_tsne_metadata(ex_sc_example, color_by = "UMI_sum", title = "UMI_sum across clusters", facet_by = "Cluster", ncol = 3)
}
