% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rl_network.R
\name{plot_rl_network}
\alias{plot_rl_network}
\title{Identifies all R / L interactions}
\usage{
plot_rl_network(input, group_by = FALSE, mode = "Summary")
}
\arguments{
\item{input}{the input full_network from calc_rc_network}

\item{group_by}{the pData columns calc_rl_network was calculated on to split the networks}

\item{mode}{the network plot type
into independent networks}
}
\description{
This function will map all RL interactions
}
\details{
This will use the calc_agg_bulk results to ID networks
}
\examples{
ex_sc_example <- id_rl(input = ex_sc_example)
}
