% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rl_summary.R
\name{plot_rl_summary}
\alias{plot_rl_summary}
\title{Identifies all R / L interactions}
\usage{
plot_rl_summary(input, group_by = FALSE,
  edge_weight = "fraction_connections", rank = TRUE, threshold = NA,
  reference = NA)
}
\arguments{
\item{input}{the input full_network from calc_rc_network}

\item{group_by}{the pData columns calc_rl_network was calculated on to split the networks}

\item{edge_weight}{the network edge_weight types. See column names of calc_rl_network$Summary}

\item{rank}{whether or not to rank the edges for contrast on arrow sizes}

\item{threshold}{a % change (as decimal ie .25 for 25%) above / below which to color, otherwise gray, valid only for grouped data}

\item{reference}{the reference on which to calculate % changes
into independent networks}
}
\description{
This function will map all RL interactions
}
\details{
This will use the calc_agg_bulk results to ID networks
}
\examples{
ex_sc_example <- id_rl(input = ex_sc_example)
}
