% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tsne_gene.R
\name{plot_tsne_gene}
\alias{plot_tsne_gene}
\title{tSNE Plot on a gene or genes}
\usage{
plot_tsne_gene(input, genes, title, colors_points = c("white", "blue", "red",
  "yellow"), colors_density = c("black", "blue", "red", "yellow"), size = 1,
  ncol = 2, cutoff = 0.2)
}
\arguments{
\item{input}{The input data}

\item{genes}{if provided will color_by the gene}

\item{title}{The title}

\item{colors_points}{Colors for the cells}

\item{colors_density}{Colors for the background}

\item{size}{The size of the points}
}
\description{
This will plot gene information onto a 2d tsne plot
}
\details{
Utilize information stored in pData to control the plot display.
}
\examples{
plot_tsne_metadata(input = ex_sc_example, genes,  title = "UMI_sum across clusters", facet_by = "Cluster", ncol = 3)
}
