% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tsne_metadata.R
\name{plot_tsne_metadata}
\alias{plot_tsne_metadata}
\title{tSNE Plot on metadata}
\usage{
plot_tsne_metadata(input, title, color_by, size = 1, colors = "NA",
  facet_by = "NA", ncol = "NA")
}
\arguments{
\item{input}{The input data}

\item{title}{The title}

\item{color_by}{What to color points by, either "UMI_sum" or pData categorial variable}

\item{colors}{What colors to utilize for categorial data. Be sure it is of the proper length!}

\item{facet_by}{What to break the plots by}

\item{ncol}{How many columns if faceting}
}
\description{
This will plot information onto a 2d tsne plot
}
\details{
Utilize information stored in pData to control the plot display.
}
\examples{
plot_tsne_metadata(input, color_by = "UMI_sum", title = "UMI_sum across clusters", facet_by = "Cluster", ncol = 3)
}
