% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tsne_multigene.R
\name{plot_tsne_multigene}
\alias{plot_tsne_multigene}
\title{tSNE Plot with multiple genes}
\usage{
plot_tsne_multigene(input, title, genes, ncol = 4, size = 1,
  colors = "NA")
}
\arguments{
\item{input}{The input data}

\item{title}{The title}

\item{genes}{if provided will color_by the gene}

\item{ncol}{How many columns if faceting}

\item{size}{The size of the points}

\item{colors}{What colors to utilize for categorial data. Be sure it is of the proper length!}
}
\description{
This will plot genes onto a 2d tsne plot
}
\details{
Utilize information stored in pData to control the plot display.
}
\examples{
plot_tsne_multigene(input, title = "UMI_sum across clusters", genes = c("Actb", "Gapdh"), ncol = 4, size = 2)
}
