% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_violin.R
\name{plot_violin}
\alias{plot_violin}
\title{This will create a violin plot}
\usage{
plot_violin(input, title = "", color_by, gene, facet_by = "NA",
  ncol = "NA", size = 1, colors = "NA", plot_mean = TRUE,
  theme = "classic")
}
\arguments{
\item{input}{The input data}

\item{title}{The title}

\item{color_by}{a pData variable}

\item{gene}{if provided will color_by the gene}

\item{facet_by}{a pData variable}

\item{ncol}{How many columns if faceting}

\item{size}{The size of the points}

\item{colors}{What colors to utilize for categorial data. Be sure it is of the proper length!}

\item{plot_mean}{If true will create a secondary axis to plot the mean expression value}
}
\description{
This will plot a given gene via violin plot
}
\details{
Utilize information stored in pData to control the plot display.
}
\examples{
plot_violin(input, title = "Actb across clusters", gene = "Actb", color_by = "Timepoint", facet_by = "Cluster", size = 1, ncol = 3)
}
