% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_violin2.R
\name{plot_violin2}
\alias{plot_violin2}
\title{plot_violin2}
\usage{
plot_violin2(
  input,
  title = "",
  gene,
  color_by,
  log_scale = F,
  colors = NA,
  facet_by = NA,
  spread = NA,
  jitter_pts = T,
  plot_mean = T,
  size = 1,
  sig = 3,
  number_labels = T,
  text_sizes = c(20, 10, 5, 10, 5, 5),
  alpha = 0.5,
  theme = "classic"
)
}
\arguments{
\item{input}{Bioconductor’s ExpressionSet Class with unlogged values in exprs().}

\item{title}{Title of the graph. Would be the gene name if not specificed.}

\item{gene}{to plot the expression level of.}

\item{color_by}{a pData variable.}

\item{log_scale}{If true, transform UMIs by log2(UMI + 1).}

\item{colors}{What colors to utilize for categorial data. Be sure it is of the proper length.}

\item{facet_by}{a vector with one or two pData variables. If two, the first variable as columns and the second as rows.}

\item{spread}{e.g. Healthy catagory is unique in Disease and Skin. To use Healthy only as skin but not Disease, that is adding Healthy skin to each disease, spread = c("Disease", "Healthy").}

\item{plot_mean}{plot the mean value as black dot with second y-axis on the right.}

\item{size}{the size of dots.}

\item{sig}{the number of digits after the decimal point for cell fraction value.}

\item{number_labels}{show the total cell numbers and cell fraction with non-zero expression values under each bar.}

\item{text_sizes}{a vector of title_size, axis_title, axis_text, legend_title, legend_text, facet_text, faults too c(20,10,5,10,5,5)}

\item{theme}{the plot theme. Default to be "classic" if not set to "bw".}
}
\description{
Create violin plot.
}
\details{
Utilize information stored in pData to control the plot display. Each point_by as a dot with a bar showing the weighted mean of all point_by dots.
}
\examples{
plot_violin2(ex_sc, gene = "CXCL13", color_by = "Skin", facet_by = c("Disease", "CellType"), log_scale = T)
plot_violin2(ex_sc, gene = "CXCL13", color_by = "Skin", facet_by = c("CellType", "Disease"), spread = T, log_scale = T)
}
