% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volcano.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Calculate Library Size}
\usage{
plot_volcano(de_path, de_file, title = NA, fdr_cut = 0.001,
  logfc_cut = 0.75)
}
\arguments{
\item{input}{the input DIRECTORY}
}
\description{
This function will count the number of reads / UMIs per cell
}
\details{
This will calculate the total number of UMIs on a per cell basis.
}
\examples{
ex_sc_example <- calc_libsize(input = ex_sc_example)
}
