% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_filter.R
\name{pre_filter}
\alias{pre_filter}
\title{Filter Cells based on Expression}
\usage{
pre_filter(input, minUMI, maxUMI = NA, threshold = NA, minCells = NA)
}
\arguments{
\item{input}{the input data}

\item{minUMI}{min number of UMIs per cell}

\item{maxUMI}{max number of UMIs per cell}

\item{threshold}{UMI threshold for gene detection}

\item{minCells}{number of cells expressed above threshold for a given gene}
}
\description{
This function will filter based on min number of UMIs, max number of UMIs, and genes based on expression across all cells.
}
\details{
When processing the data, low quality cells may contain very few UMIs, while some overrepresented cell barcodes may indicate barcode bleedover or celll doublets. Filtering out both low and high UMI count cells is recommended before normalization.
}
\examples{
filtered_data <- filter_UMIs(input = ex_sc_example, minUMI = 1000, maxUMI = 10000, threshold = 1, minCells = 10)
}
