% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Receptor_Ligand_Data}
\alias{Receptor_Ligand_Data}
\title{Receptor Ligand Data}
\format{A data frame with 2557 rows and 5 variables:
\itemize{
  \item Pair.Name: A specific receptor ligand pair
  \item Ligand.ApprovedSymbol: the gene name
  \item Ligand.Name: common name of the ligand
  \item Receptor.ApprovedSymbol: the gene name
  \item Receptor.Name: common name of the receptor
}}
\usage{
data(Receptor_Ligand_Data)
}
\description{
Simplified Receptor Ligand database
}
\references{
Ramilowski, Jordan A., et al. "A draft network of ligand–receptor-mediated multicellular signalling in human." Nature communications 6 (2015): 7866.

A dataset containing a simplified version of the datafile in reference above.
}
\keyword{datasets}
