% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_agg_gene.R
\name{return_agg_gene}
\alias{return_agg_gene}
\title{Normalization}
\usage{
return_agg_gene(input)
}
\arguments{
\item{input}{the input ex_sc}
}
\description{
This function will perform normalization of your data
}
\details{
If the method is ICA, independent component analysis will be performed, and then tSNE will do the final dimension reduction. If PCA is selected, PCA will be performed before on the expression matrix transpose before tSNE. This PCA will use the cells positions on the principal components. If iPCA is selected, PCA will be be performed but without transposing the data. This will create "meta cells" instead of meta genes created in the typical PCA. Then tSNE will be performed on each cells contribution (loading) to the meta cell. We find that iPCA is much more robust and leads to cleaner clusters than traditional PCA.
}
\examples{
ex_sc_example <- dim_reduce(input = ex_sc_example, genelist = gene_subset, pre_reduce = "iPCA", nComp = 15, tSNE_perp = 30, iterations = 500, print_progress=TRUE)

}
