% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_markers.R
\name{return_markers}
\alias{return_markers}
\title{Return markers}
\usage{
return_markers(input)
}
\arguments{
\item{input}{the input data matrix.}
}
\description{
This will return the marker genes.
}
\details{
This will return the marker and reference genes stored in fData(input) in a convienient list format.
}
\examples{
marker_list <- return_markers(input)
}
