% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_ggplot.R
\name{save_ggplot}
\alias{save_ggplot}
\title{Save plot}
\usage{
save_ggplot(filename = "Rplot", format = "pdf", dingbats = FALSE,
  dpi = 300, h = 6, w = 6, units = "in")
}
\arguments{
\item{filename}{The file name to be written}

\item{format}{File extension can be "pdf", "svg", "tiff", "png", "jpg"}

\item{dingbats}{FALSE or TRUE (for pdf only)}

\item{dpi}{resolution for raster output types}

\item{h}{The height}

\item{w}{The Width}

\item{units}{"in", "cm"}
}
\description{
This will save the last ggplot call
}
\details{
This will save the save ggplot object that was written to the device.
}
\examples{
save_plot(filename = "Cluster plot", format = "pdf")
}
