% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_gene.R
\name{search_gene}
\alias{search_gene}
\title{Search Gene}
\usage{
search_gene(input, search)
}
\arguments{
\item{input}{the input ex_sc}

\item{search}{The keyword or partial word to match}
}
\description{
This function will search for a gene in your data by some input
}
\details{
This will calculate total UMIs on a per cell basis.
}
\examples{
ex_sc_example <- calc_libsize(input = ex_sc_example)
}
