% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_genes.R
\name{subset_genes}
\alias{subset_genes}
\title{Select Genes}
\usage{
subset_genes(input, threshold, minCells)
}
\arguments{
\item{input}{the input data matrix.}

\item{threshold}{UMI threshold for gene detection}

\item{minCells}{number of cells expressed above threshold for a given gene}
}
\description{
This will select genes based on minimum expression and coefficient of variation, or by a preliminary PCA that selects genes based on malhanobis distance from the center.
}
\details{
This selects genes.
}
\examples{
gene_subset <- subset_genes(exprs(ex_sc_example), threshold = 1, minCells = 10)
}
