% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_rl_network.R
\name{analyze_rl_network}
\alias{analyze_rl_network}
\title{Identifies all R / L interactions}
\usage{
analyze_rl_network(input, h = 8, w = 8, prefix = "", mult = 1,
  layout = FALSE, subset = FALSE, subset_on = "connected")
}
\arguments{
\item{input}{the input network from plot_rl_network}

\item{h}{the height of images}

\item{w}{the height of images}

\item{prefix}{a character to be appended to start of file names}
}
\description{
This function will map all RL interactions
}
\details{
This will use the calc_agg_bulk results to ID networks
}
\examples{
ex_sc_example <- id_rl(input = ex_sc_example)
}
