% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_sc.R
\name{cluster_sc}
\alias{cluster_sc}
\title{Cluster_sc
This will perform clustering on your single cell data.}
\usage{
cluster_sc(input, dimension, method, num_clust)
}
\arguments{
\item{input}{the input data matrix.}

\item{dimension}{either "Comp" or "2d"}

\item{method}{can either be "spectral" or "density"}

\item{num_clust}{the number of clusters}
}
\description{
Cluster_sc
This will perform clustering on your single cell data.
}
\details{
This will perform clustering on either the high dimensional PCA / ICA components or on the final 2d result from tsne
}
\examples{
ex_sc_example <- cluster_sc(input = ex_sc_example, dimension = "Comp", method = "spectral", num_clust = 6)
}
