% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_reduce.R
\name{dim_reduce}
\alias{dim_reduce}
\title{Initial Dimension Reduction}
\usage{
dim_reduce(input, threshold, minCells, nComp_ICA, tSNE_perp,
  print_progress = FALSE)
}
\arguments{
\item{input}{the input data matrix}

\item{threshold}{UMI threshold for gene detection}

\item{minCells}{number of cells expressed above threshold for a given gene}

\item{nComp_ICA}{the number of ICA components to reduce too, 5-20 recommended}

\item{tSNE_perp}{number of cells expressed above threshold for a given gene, 10-100 recommended}
}
\description{
This function will do a preliminary dimension reduction. This is necessary because normalization across cell types requires preliminary clusters.
}
\details{
Before normalization preliminary clustering is needed in order to normalize internally within a cluster before normalizing across clusters. The minCells will be highly dependent on the number of cells in your dataset. In general aim to include genes expressed in no less than 1:100 cells, unless you have extremely rare cell types below that threshold.
}
\examples{
filtered_data <- initial_dim_reduce(input = filtered_data, threshold = 5, minCells = 50, nComp_ICA = 10, tSNE_perp = 30, print_progress=TRUE)
}
