% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_rl_network.R
\name{filter_rl_network}
\alias{filter_rl_network}
\title{Identifies all R / L interactions}
\usage{
filter_rl_network(input, filter_by, filter_type = "network",
  DEfolder = NULL, cutoff, direction, absol = FALSE, group_by = FALSE,
  keep = FALSE)
}
\arguments{
\item{input}{the output of calc_rl_network}

\item{filter_by}{the column to summarize by}

\item{filter_type}{"network" or "DE"}

\item{DEfolder}{the output location of findDEgenes}

\item{cutoff}{value to filter by}

\item{direction}{>, <, =}

\item{absol}{whether or not to take the absolute value of the column}

\item{keep}{if true will retain all edges and annotate them as passing the filter or not}
}
\description{
This function will map all RL interactions
}
\details{
This will use the calc_agg_bulk results to ID networks
}
\examples{
ex_sc_example <- id_rl(input = ex_sc_example)
}
