% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findDEmarkers.R
\name{findDEmarkers}
\alias{findDEmarkers}
\title{This will perform differential expression using edgeR to find markers for each cluster or group compared to all others}
\usage{
findDEmarkers(
  input,
  pd = NULL,
  DEgroup,
  sizefactor = NULL,
  lib_size = NULL,
  batchID = NULL,
  outdir = "DEmarkers/",
  outsuffix = "DEmarkers.tsv",
  minCells = 0.01,
  pVal = 1,
  contrast = list(c(-1, 1))
)
}
\arguments{
\item{input}{Input expression set}

\item{pd}{pData, optional, by default set to pData of input}

\item{DEgroup}{a pData variable}

\item{sizefactor}{a pData column containing the scran reported size factor for each cell}

\item{lib_size}{a pData column containing the library size for each cell}

\item{batchID}{a pData variable to use for model accounting for batch effects}

\item{outdir}{an output directory}

\item{outsuffix}{an output file suffix}

\item{minCells}{minimum cell fraction the genes should be expressed in to be tested for DE}

\item{pVal}{pvalue cutoff for reported results, by default reports all results}

\item{contrast}{a list of contrasts to be used for the DE analysis default is a two-class comparison, with the second class as the comparison class}
}
\description{
This will perform differential expression using edgeR to find markers for each cluster or group compared to all others
}
\details{
Utilize information stored in pData to control the DE performed
}
\examples{
findDEmarkers(input, pd = selected_table, DEgroup = "cluster", batchID = "beads", sizefactor="sizefactor", lib_size="lib_size")
}
