% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_markers.R
\name{id_markers}
\alias{id_markers}
\title{ID markers}
\usage{
id_markers(input, id_by = "Cluster", print_progress = TRUE, overwrite = FALSE)
}
\arguments{
\item{input}{the input ex_sc}

\item{id_by}{a pData variable to run id_markers on. Defaults to "Cluster"}

\item{print_progress}{will print progress if TRUE}

\item{overwrite}{if TRUE will overwrite previously calculated scores on the same id_by variable}
}
\description{
This will perform marker gene identification
}
\details{
This will find marker genes for each cluster. First, for each cluster it calculates the fraction of cells expressing a given gene,
and the mean expression of that gene (within the cluster).
It then compares these values to the values of all other groups looking for genes which maximize the distance in both
fraction expressing and the mean expression.
}
\examples{
ex_sc_example <- id_markers(input = ex_sc_example, print_progress = TRUE)
}
