% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_density.R
\name{plot_density}
\alias{plot_density}
\title{Plot Density}
\usage{
plot_density(input, val, color_by = "NA", statistic = "mean")
}
\arguments{
\item{input}{the input data}

\item{val}{Either "UMI_sum" or a gene name}

\item{color_by}{The metadata in pData that will be used to break the distributions into groups. Can be left blank.}

\item{statistic}{Either the "mean" or "median" can be calculated}
}
\description{
This function will draw a distribution for a given value
}
\details{
This will draw a UMI distribution for each variable in the metadat.
}
\examples{
draw_density(input = ex_sc_example, val = "UMI_sum", color_by = "Cluster", statistic = "mean")
}
