% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_go_heatmap.R
\name{plot_go_heatmap}
\alias{plot_go_heatmap}
\title{Construct Expression Set Class}
\usage{
plot_go_heatmap(input, sig_val = "p.adjust", cutoff = 0.1,
  max_categories = 30, subset_group = F)
}
\arguments{
\item{input}{the input Go result from clusterProfiler}

\item{sig_val}{p.adjust of pvalue}

\item{cutoff}{the value to be included}

\item{max_categories}{mox categories included}

\item{subset_group}{table(input@compareClusterResult$group)}
}
\description{
This function will take an input expression matrix and make an Expression Set Class
}
\details{
This will take an input matrix and create the expression set class that further analysis
will be written to
}
\examples{
construct_ex_sc(input = sc_dat)
}
