% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rl_network.R
\name{plot_rl_network}
\alias{plot_rl_network}
\title{Identifies all R / L interactions}
\usage{
plot_rl_network(input, input_full, group_by = FALSE, comparitive = FALSE,
  from = FALSE, to = FALSE, value = FALSE, layout = "nicely",
  write_interactive = TRUE, interactive_groups = "nodes", nodesize = 3,
  size_by_connections = TRUE, textsize = 0.5, h = 8, w = 8,
  prefix = "")
}
\arguments{
\item{input}{the input networked (usually filtered)}

\item{input_full}{the input full network}

\item{group_by}{the pData columns calc_rl_network was calculated on to split the networks}

\item{comparitive}{If true will look at connections across group_by and color edges by their change}

\item{from}{the comparitive reference}

\item{to}{the condition to compare to reference}

\item{value}{the column of the full network to calculate foldChange on}

\item{layout}{nicely, kk, circle}

\item{write_interactive}{whether or not to write an interactive visNetwork html object}

\item{interactive_groups}{the dropdown menu for selection nodes, either "nodes", "group_by", or "cluster"}

\item{nodesize}{The size of nodes}

\item{size_by_connections}{if true will override node size and size by the degree of the node}

\item{textsize}{The size of text}

\item{h}{pdf height}

\item{w}{pdf width}

\item{prefix}{a character to be appeneded to the start of file names
into independent networks}
}
\description{
This function will map all RL interactions
}
\details{
This will use the calc_agg_bulk results to ID networks
}
\examples{
ex_sc_example <- id_rl(input = ex_sc_example)
}
