% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rl_network.R
\name{plot_rl_network}
\alias{plot_rl_network}
\title{Identifies all R / L interactions}
\usage{
plot_rl_network(input, break_by = FALSE)
}
\arguments{
\item{input}{the input ex_sc}

\item{break_by}{the pData columns calc_agg_bulk was calculated on}
}
\description{
This function will map all RL interactions
}
\details{
This will use the calc_agg_bulk results to ID networks
}
\examples{
ex_sc_example <- id_rl(input = ex_sc_example)
}
