% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tsne_gene.R
\name{plot_tsne_gene}
\alias{plot_tsne_gene}
\title{tSNE Plot on a gene or gene}
\usage{
plot_tsne_gene(input, gene, title = "", density = FALSE, facet_by = "NA",
  colors_points = c("gray", "blue", "red", "yellow"), size = 1.5,
  ncol = 2, resolution = 500, cutoff = 0.2)
}
\arguments{
\item{input}{The input data}

\item{gene}{the gene or gene to ploy}

\item{title}{The title}

\item{density}{If true will color a density of the points}

\item{facet_by}{If ONE gene is being plotted it can be faceted. You cannot facet multiple gene}

\item{colors_points}{Colors for the cells}

\item{size}{The size of the points}

\item{ncol}{controls the number of columns if faceting}

\item{resolution}{Control the density drawing resolution. Higher values will increase file size}

\item{cutoff}{removes points from density below this threshold. Lower values will increase file size}
}
\description{
This will plot gene information onto a 2d tsne plot
}
\details{
Utilize information stored in pData to control the plot display.
}
\examples{
plot_tsne_gene(ex_sc_example, gene = "Tnf", title = "Tnf over Time", facet_by = "Timepoint", density = TRUE)

}
