% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotViolin.R
\name{plotViolin}
\alias{plotViolin}
\title{This will create a violin plot}
\usage{
plotViolin(
  gene,
  input,
  sampleID,
  colourID,
  facetID = NULL,
  cols = NULL,
  subsetID = NA,
  subsetName = NA,
  facet_scale = "fixed",
  mean_text = F,
  fraction_text = F,
  type = "violin",
  fudge = 0
)
}
\arguments{
\item{gene}{Will only plot data for this gene, if the gene paramter is a list the violin plot will display the sum of the expression for all genes in the list}

\item{input}{Input expression set}

\item{sampleID}{pData variable to group by on the x axis}

\item{colourID}{a pData variable to color by}

\item{facetID}{an optional pData variable to plot by}

\item{cols}{Personalized colour vector, length should equal number of groups in colourID}

\item{subsetID}{a pData variable to subset the original data by}

\item{subsetName}{a value from the subsetID variable specified, this will select only these cells to plot}

\item{facet_scale}{ggplot2 parameter for plot scales defaults to fixed}

\item{mean_text}{ggplot2 parameter for adding the mean expression per group defaults to False}

\item{fraction_text}{ggplot2 parameter for adding the fraction of cells with expression greater than 0 for that gene defaults to False}

\item{type}{plot type, default is violin, can also plot density and cdf distribution}

\item{fudge}{a pseudocount value to avoid dropping cells with zero counts}
}
\description{
This will plot a given gene via violin plot
}
\details{
Utilize information stored in pData to control the plot display.
}
\examples{
plotViolin("Actb", input, sampleID = "sample", facetID = "subtype", colourID = "genotype", cols = NULL, subsetID = "celltype", subsetName = "Neurons", facet_scale = "fixed")
}
