% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_filter.R
\name{pre_filter}
\alias{pre_filter}
\title{Filter Data}
\usage{
pre_filter(
  input,
  minUMI = 0,
  maxUMI = NA,
  threshold = NA,
  minCells = NA,
  print_progress = TRUE
)
}
\arguments{
\item{input}{the input ex_sc}

\item{minUMI}{min number of UMIs per cell}

\item{maxUMI}{max number of UMIs per cell}

\item{threshold}{UMI threshold for gene detection}

\item{minCells}{number of cells expressed above threshold for a given gene}

\item{print_progress}{will print progress if TRUE}
}
\description{
This function will filter the data to remove cells and/or genes
}
\details{
When processing the data, low quality cells may contain very few UMIs,
while some overrepresented cell barcodes may indicate barcode bleedover or cell doublets.
Filtering out both low and high UMI count cells is recommended before normalization.
}
\examples{
filtered_data <- filter_UMIs(input = ex_sc_example, minUMI = 1000, maxUMI = 10000, threshold = 1, minCells = 10, print_progress = TRUE)
}
