% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_markers.R
\name{return_markers}
\alias{return_markers}
\title{Return markers}
\usage{
return_markers(input, num_markers)
}
\arguments{
\item{input}{the input ex_sc}

\item{num_markers}{the number of markers to return for each cell type}
}
\description{
This will return the marker genes.
}
\details{
This will return the marker and reference genes stored in fData(input) in a convienient list format.
}
\examples{
marker_list <- return_markers(input = ex_sc, num_markers = 50)
}
