% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_genes.R
\name{subset_genes}
\alias{subset_genes}
\title{Select Genes}
\usage{
subset_genes(input, method, threshold, minCells, nComp, cutoff)
}
\arguments{
\item{input}{the input data matrix.}

\item{method}{can either be "Expression", CV", or "PCA"}

\item{threshold}{UMI threshold for gene detection}

\item{minCells}{number of cells expressed above threshold for a given gene}

\item{nComp}{if method = PCA, the number of components to keep}

\item{cutoff}{the percentile of genes by coefficient of variation or PCA loadings to keep}
}
\description{
This will select genes based on minimum expression and coefficient of variation, or by a preliminary PCA that selects genes based on malhanobis distance from the center.
}
\details{
This selects genes.
}
\examples{
gene_subset <- subset_genes(input = exprs(ex_sc_example), method = "PCA", threshold = 3, minCells = 30, nComp = 15, cutoff = 0.75)
}
